windows: clean type warnings
There are some minor public api type improvements rather than cast everywhere
inside lws and user code to work around them... these changed from int to
size_t

 - lws_buflist_use_segment() return
 - lws_tokenize_t .len and .token_len
 - lws_tokenize_cstr() length
 - lws_get_peer_simple() namelen
 - lws_get_peer_simple_fd() namelen, int fd -> lws_sockfd_type fd
 - lws_write_numeric_address() len
 - lws_sa46_write_numeric_address() len

These changes are typically a NOP for user code

+++ lib/core-net/connect.c
@@ -239,7 +239,7 @@ lws_client_connect_via_info(const struct lws_client_connect_info *i)
 	for (n = 0; n < CIS_COUNT; n++)
 		if (cisin[n]) {
 			wsi->stash->cis[n] = pc;
-			m = strlen(cisin[n]) + 1;
+			m = (int)strlen(cisin[n]) + 1;
 			memcpy(pc, cisin[n], m);
 			pc += m;
 		}
